wip: redo network structure
