+++ src/earthkit/hydro/readers/__init__.py
@@ -1,0 +1,6 @@
+from .readers import (
+    find_main_var,
+    from_cama_nextxy,
+    from_d8,
+    import_earthkit_or_prompt_install,
+)
