+++ src/earthkit/hydro/readers/_numba.py
@@ -8,7 +8,6 @@ def get_edge_indices_numba(offsets, grouping):
     total_len = np.sum(lengths)
     result = np.empty(total_len, dtype=np.int64)
     pos = 0
-
     for i in range(len(grouping)):
         node = grouping[i]
         length = lengths[i]
