+++ src/earthkit/hydro/readers/network_utils.py
