+++ src/earthkit/hydro/readers/topological_labels_rust.py
@@ -8,7 +8,7 @@
 
 import numpy as np
 
-from ._rust import propagate_labels
+from earthkit.hydro._rust import propagate_labels
 
 
 def compute_topological_labels(
@@ -33,8 +33,8 @@ def compute_topological_labels(
     """
 
     n_nodes = np.uintp(downstream_nodes.shape[0])
-    labels = np.zeros(n_nodes, dtype=np.int64)
+    # labels = np.zeros(n_nodes, dtype=np.int64)
 
-    labels = propagate_labels(labels, sources, sinks, downstream_nodes, n_nodes)
+    labels = propagate_labels(sources, sinks, downstream_nodes, n_nodes)
 
     return labels
